requirements.txt
africastalking==1.1.5
apns2==0.5.0
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
backcall==0.1.0
black==19.3b0
boto3==1.9.111
botocore==1.12.111
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
Click==7.0
contextlib2==0.5.5
cryptography==2.7
decorator==4.4.0
Django==2.1.5
django-appconf==1.0.3
django-colorfield==0.1.15
django-cryptography==0.3
django-dotenv==1.4.2
django-filter==2.1.0
django-push-notifications==1.6.1
django-storages==1.7.1
django-stubs==1.1.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.2.0
docutils==0.14
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.8
ipython==7.3.0
ipython-genutils==0.2.0
isort==4.3.15
jedi==0.13.3
jmespath==0.9.4
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
mypy==0.720
mypy-extensions==0.4.1
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.9
psycopg2==2.7.7
ptyprocess==0.6.0
py-vapid==1.7.0
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
pylint==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2018.9
pywebpush==1.10.0
requests==2.21.0
s3transfer==0.2.0
schema==0.7.0
six==1.12.0
toml==0.10.0
traitlets==4.3.2
typed-ast==1.3.1
typing-extensions==3.7.4
urllib3==1.24.1
wcwidth==0.1.7
wrapt==1.11.1
